Totally rewrite random_chests.lua
This commit totally rewrites and reworks several aspects of
the random_chests.lua file. A summary of the changes is as
follows.

- Simplify and improve configuration
- Totally rework chest generation code making it much more readable
and allowing the user to specify node groups which chests will not
spawn on
- Refactor code to follow minetest lua style conventions
- Store configuration variables in the random_chests table, rather
than in a local variable accessed via a setter function
- Remove the group parameter from random_chests.register_item
as it was unnecessary